+++ packages/fcl/src/current-user/exec-service/plugins.js
@@ -1,6 +1,6 @@
 import {invariant} from "@onflow/util-invariant"
 import {LEVELS, log} from "@onflow/util-logger"
-import {isRequired, isString, isObject, isFunc} from "../../exec/utils/is"
+import {isRequired, isString, isObject, isFunc} from "../../utils/is"
 import { CORE_STRATEGIES } from "../../utils/constants"
 
 const stub = () => {
