+++ packages/fcl/src/exec/mutate.js
@@ -3,7 +3,7 @@ import {normalizeArgs} from "./utils/normalize-args"
 import {getCurrentUser} from "../current-user"
 import {prepTemplateOpts} from "./utils/prep-template-opts.js"
 import {preMutate} from "./utils/pre.js"
-import {isNumber} from "./utils/is"
+import {isNumber} from "../utils/is"
 
 export const getMutate = ({platform}) => {
   /**
