+++ packages/fcl/src/utils/is.js
